requirements.txt
click>=8.1.7
colorama>=0.4.6
coqui-tts>=0.24.1
numpy>=1.26.4
ollama>=0.2.1
PyAudio>=0.2.14
pydantic-settings>=2.3.1
pygame>=2.5.2
torch>=2.3.1
torchaudio>=2.3.1
transformers>=4.40.2


fastapi
uvicorn
python-multipart
pydub
